refactor(discussion/arena): port TOC, remove legacy viewer, optimistic top-level like
- Port TOC to arena (sticky right column on desktop when ≥ 3 headings)
- Remove `views/main/` subtree + `components/discussion_comments.rs` +
  `discussion_comment_context.rs`; editor route retained
- Reroute `SpaceActionSummary::get_url()` TopicDiscussion branch and
  prerequisite_card off the removed viewer route
- Mirror reply-like optimistic update on top-level comment likes
- Drop page-entries test for removed discussion viewer URL
