--- app/ratel/src/features/spaces/pages/actions/actions/discussion/components/mod.rs
@@ -1,3 +1,0 @@
-mod discussion_comments;
-
-pub use discussion_comments::DiscussionComments;
