+++ app/ratel/src/features/spaces/pages/actions/actions/discussion/mod.rs
@@ -1,21 +1,16 @@
-mod components;
 mod config;
 mod context;
-mod discussion_comment_context;
 mod models;
 mod types;
 mod views;
 
 pub mod controllers;
 
-use components::*;
 use context::*;
 use controllers::*;
-use discussion_comment_context::*;
 use models::*;
 
 pub use context::*;
-pub use discussion_comment_context::*;
 pub use models::*;
 pub use types::*;
