--- app/ratel/src/features/spaces/pages/actions/actions/discussion/views/main/creator/i18n.rs
@@ -1,82 +1,0 @@
-use crate::features::spaces::pages::actions::actions::discussion::*;
-
-translate! {
-    DiscussionCreatorTranslate;
-
-    page_title: {
-        en: "Discussion",
-        ko: "토론",
-    },
-    tab_overview: {
-        en: "Overview",
-        ko: "개요",
-    },
-    tab_upload: {
-        en: "Upload",
-        ko: "업로드",
-    },
-    tab_setting: {
-        en: "Settings",
-        ko: "설정",
-    },
-    upload_drop_title: {
-        en: "Drag & Drop files here",
-        ko: "여기에 파일을 드래그 앤 드롭하세요",
-    },
-    upload_cta: {
-        en: "Upload",
-        ko: "업로드",
-    },
-    upload_supported_types: {
-        en: "PDF, DOCX, PPTX, XLSX, PNG, JPG, MP4, MOV 100MB",
-        ko: "PDF, DOCX, PPTX, XLSX, PNG, JPG, MP4, MOV 100MB",
-    },
-    upload_empty: {
-        en: "No uploaded files yet.",
-        ko: "아직 업로드된 파일이 없습니다.",
-    },
-    upload_view: {
-        en: "View",
-        ko: "보기",
-    },
-    upload_delete: {
-        en: "Delete",
-        ko: "삭제",
-    },
-    title_label: {
-        en: "Title",
-        ko: "제목",
-    },
-    title_placeholder: {
-        en: "Enter discussion title...",
-        ko: "토론 제목을 입력하세요...",
-    },
-    description_placeholder: {
-        en: "Enter discussion description...",
-        ko: "토론 설명을 입력하세요...",
-    },
-    category_label: {
-        en: "Category",
-        ko: "카테고리",
-    },
-    category_placeholder: {
-        en: "Enter category (optional)...",
-        ko: "카테고리를 입력하세요 (선택)...",
-    },
-    btn_save: {
-        en: "Save",
-        ko: "저장",
-    },
-    saving: {
-        en: "Saving...",
-        ko: "저장 중...",
-    },
-    all_changes_saved: {
-        en: "All changes saved",
-        ko: "모든 변경사항 저장됨",
-    },
-    unsaved_changes: {
-        en: "Unsaved changes",
-        ko: "저장되지 않은 변경사항",
-    },
-}
