--- app/ratel/src/features/spaces/pages/actions/actions/discussion/views/main/viewer/i18n.rs
@@ -1,15 +1,0 @@
-use dioxus_translate::*;
-
-translate! {
-    DiscussionViewerTranslate;
-
-    untitled_discussion: { en: "Untitled Discussion", ko: "제목 없는 토론" },
-    comments: { en: "Comments", ko: "댓글" },
-    comments_count: { en: "comments", ko: "개의 댓글" },
-
-    attachments: { en: "Attachments", ko: "첨부파일" },
-    table_of_contents: { en: "On this page", ko: "목차" },
-
-    open_comments: { en: "Open comments", ko: "댓글 열기" },
-    close_comments: { en: "Close comments", ko: "댓글 닫기" },
-}
