+++ app/ratel/src/features/spaces/pages/actions/actions/discussion/views/mod.rs
@@ -1,6 +1,3 @@
-mod main;
-pub use main::DiscussionActionPage;
-
 mod editor;
 pub use editor::DiscussionActionEditorPage;
