+++ app/ratel/src/features/spaces/pages/index/action_pages/discussion/mod.rs
@@ -1,5 +1,7 @@
 mod component;
 mod i18n;
+mod toc;
 
 pub use component::*;
 use i18n::*;
+pub use toc::*;
